feat: added test cases for mistral model support
